refactoring: migrate from unittest to pytest
