+++ requirements.in
@@ -1,1 +1,2 @@
 requests
+pytest
